pyproject.toml
[project]
name = "erp3"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.0",
    "psycopg2>=2.9.10",
]
